+++ src/db/storage/sqlite.rs
@@ -1,8 +1,8 @@
 use crate::db::encryption::Encryption;
+use crate::db::models::record::AccountName;
 use crate::db::models::secure_record::SecureRecord;
 use crate::db::storage::StorageTrait;
 use crate::db::Connection;
-use crate::storage::accounts::AccountName;
 use crate::{Db, Record, TotpError};
 use r2d2_sqlite::rusqlite::params;
 use std::collections::HashMap;
@@ -26,15 +26,23 @@ impl SqliteStorage {
 }
 
 impl StorageTrait for SqliteStorage {
-    fn search_account(&self, account_search: String) -> Result<(AccountName, Record), TotpError> {
-        todo!()
-    }
+    fn search_accounts(&self, account_search: &str) -> Result<Vec<Record>, TotpError> {
+        let accounts = self.accounts()?;
+        let mut accounts = accounts
+            .iter()
+            .filter(|(account_name, _secure_data)| {
+                account_name
+                    .to_lowercase()
+                    .contains(&account_search.to_lowercase())
+            })
+            .clone()
+            .collect::<Vec<(&AccountName, &Record)>>();
+        accounts.sort_by(|a, b| a.0.cmp(b.0));
 
-    fn search_accounts(
-        &self,
-        account_search: String,
-    ) -> Result<Vec<(AccountName, Record)>, TotpError> {
-        todo!()
+        Ok(accounts
+            .into_iter()
+            .map(|(_a, t)| t.clone())
+            .collect::<Vec<_>>())
     }
 
     fn add_account(&self, record: Record) -> Result<(), TotpError> {
@@ -65,18 +73,23 @@ impl StorageTrait for SqliteStorage {
         Ok(())
     }
 
-    fn remove_account(&self, account_or_id: String) -> Result<(), TotpError> {
+    fn remove_account(&mut self, account_or_id: String) -> Result<(), TotpError> {
         if let Ok(id) = account_or_id.parse::<u32>() {
             return self.remove_account_by_id(id);
         }
         self.remove_account_by_name(account_or_id)
     }
-    fn remove_account_by_name(&self, account: AccountName) -> Result<(), TotpError> {
-        todo!()
+    fn remove_account_by_name(&mut self, account: AccountName) -> Result<(), TotpError> {
+        let account = self.search_account(&account)?;
+        self.remove_account_by_id(account.id)
     }
 
-    fn remove_account_by_id(&self, id: u32) -> Result<(), TotpError> {
-        todo!()
+    fn remove_account_by_id(&mut self, id: u32) -> Result<(), TotpError> {
+        const SQL: &str = "DELETE FROM secure_records WHERE id = ?1;";
+        let conn = Connection::try_from(&self.db)?;
+        let mut stmt = conn.prepare(SQL)?;
+        stmt.execute(params![&id])?;
+        self.load()
     }
 
     fn accounts(&self) -> Result<HashMap<AccountName, Record>, TotpError> {
@@ -91,12 +104,102 @@ impl StorageTrait for SqliteStorage {
         Ok(accounts)
     }
 
-    fn password(&self) -> &str {
-        self.db.password()
-    }
     fn load(&mut self) -> Result<(), TotpError> {
         let conn = Connection::try_from(&self.db)?;
         self.secure_records = SecureRecord::all(&conn)?;
         Ok(())
+    }
+
+    fn password(&self) -> &str {
+        self.db.password()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::env::temp_dir;
+    use std::str::FromStr;
+    use std::time::{SystemTime, UNIX_EPOCH};
+
+    fn rand() -> u32 {
+        SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .subsec_nanos()
+    }
+
+    fn get_filename() -> String {
+        let mut dir = temp_dir();
+        dir.push(format!(".test.storage.{}.txt", rand()));
+        dir.to_string_lossy().to_string()
+    }
+
+    #[test]
+    fn tokens_ignore_case() {
+        let token_string = "jbswy3dpehpk3pxp";
+        let token: Token = token_string.parse().unwrap();
+        assert_eq!(token, Token::from_str("JBSWY3DPEHPK3PXP").unwrap());
+    }
+
+    #[test]
+    fn add_account() {
+        let filename = get_filename();
+        let mut storage = Storage::new("password".to_string(), Some(filename.clone())).unwrap();
+        assert_eq!(storage.to_iter().len(), 0);
+        storage
+            .add_account(
+                "Account1".to_string(),
+                Token::from_str("JBSWY3DPEHPK3PXP").unwrap(),
+            )
+            .unwrap();
+        assert_eq!(storage.to_iter().len(), 1);
+        let storage = Storage::new("password".to_string(), Some(filename.clone())).unwrap();
+        assert_eq!(storage.to_iter().len(), 1);
+        let _ = std::fs::remove_file(filename);
+    }
+
+    #[test]
+    fn delete_account() {
+        let filename = get_filename();
+        let mut storage = Storage::new("password".to_string(), Some(filename.clone())).unwrap();
+        assert_eq!(storage.to_iter().len(), 0);
+        storage
+            .add_account(
+                "Account1".to_string(),
+                Token::from_str("JBSWY3DPEHPK3PXP").unwrap(),
+            )
+            .unwrap();
+        assert_eq!(storage.to_iter().len(), 1);
+        storage
+            .add_account(
+                "Account2".to_string(),
+                Token::from_str("JBSWY3DPEHPK3PXP").unwrap(),
+            )
+            .unwrap();
+        assert_eq!(storage.to_iter().len(), 2);
+        storage.remove_account("Account1".into()).unwrap();
+        assert_eq!(storage.to_iter().len(), 1);
+        let _ = std::fs::remove_file(filename);
+    }
+
+    #[test]
+    fn get_account_token() {
+        let filename = get_filename();
+        let mut storage = Storage::new("password".to_string(), Some(filename.clone())).unwrap();
+        storage
+            .add_account(
+                "Account1".to_string(),
+                Token::from_str("JBSWY3DPEHPK3PXP").unwrap(),
+            )
+            .unwrap();
+        storage
+            .add_account("Account2".to_string(), Token::from_str("KRSXG5A=").unwrap())
+            .unwrap();
+        let (_, token) = storage.search_account("Account2".into()).unwrap();
+        assert_eq!(token.to_string(), "KRSXG5A=".to_string());
+        let token = storage.search_account("Account3".into());
+        assert!(token.is_err());
+
+        let _ = std::fs::remove_file(filename);
     }
 }
